requirements.txt
absl-py==2.0.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
ansi2html==1.8.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
black==23.10.1
blinker==1.7.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.5
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
descartes==1.1.0
einops==0.6.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
fire==0.5.0
flake8==6.1.0
Flask==3.0.0
fonttools==4.43.1
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h5py==2.10.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipython==8.12.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.6.12
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
model-index==0.1.11
mypy-extensions==1.0.0
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==2.2
numba==0.53.0
numpy==1.23.5
nuscenes-devkit==1.1.11
oauthlib==3.2.2
open3d-python
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
plyfile==1.0.1
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==3.19.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.2
rpds-py==0.10.6
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.4.1
Shapely==1.8.5.post1
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.2.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.5.4
tomli==2.0.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.65.2
traitlets==5.13.0
trimesh==2.35.39
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
waymo-open-dataset-tf-2-2-0==1.2.0
wcwidth==0.2.9
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
yapf==0.40.1
zipp==3.17.0
